# pubky-homeserver/openapi.yml
openapi: 3.1.0
info:
  title: Pubky Homeserver API
  description: |
    The Pubky Homeserver exposes three servers:
    - **Client Server** — Primary API for user authentication, file storage, and event streaming.
    - **Admin Server** — Password-protected administration endpoints.
    - **Metrics Server** — Prometheus metrics (`GET /metrics`).

    This specification covers the Client Server and Admin Server.

    ## Tenant Identification

    Per-tenant (user) routes resolve the target user from these sources (highest priority first):
    1. `pubky-host` header (explicit override; takes precedence over `Host`)
    2. `Host` header (e.g. from TLS SNI on the Pubky socket)
    3. `?pubky-host=` query parameter (fallback when no key found in headers)

    The value must be a z-base-32 encoded Ed25519 public key.

    ## Authentication

    Two authentication methods exist:
    - **Grant-based (current):** `Authorization: Bearer <token>` obtained via `POST /auth/grant/session`.
    - **Cookie-based (deprecated):** Session cookie set by `POST /signup` or `POST /session`.

    Grant middleware runs first; if it produces a session, cookie middleware is skipped.

    ## Capabilities

    Write operations are gated by capability-based access control. Capabilities are
    path-scoped (e.g. `read_write:/pub/app/`) or `root` (full access to `/pub/`).

    ## Write Path Restriction

    All PUT/DELETE operations require paths under `/pub/`. Attempts to write elsewhere
...
  version: 0.9.0
...
servers:
...
paths:
...
  /users/{pubkey}/quota:
    get:
      tags: [Admin]
      summary: Get user quota
      description: |
        Returns both the effective quota (overrides merged with system defaults)
        and the raw per-user overrides.
      operationId: getUserQuota
      servers:
        - url: http://localhost:6288
      security:
        - adminPassword: []
      parameters:
        - $ref: "#/components/parameters/AdminPubkey"
      responses:
        "200":
          description: User quota
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserQuotaResponse"
              examples:
                withDefaults:
                  value:
                    effective:
                      storage_quota_mb: 100
                      rate_read: "10mb/s"
                      rate_write: "5mb/s"
                    overrides: {}
                withOverrides:
                  value:
                    effective:
                      storage_quota_mb: 500
                      rate_read: "100mb/m"
                      rate_write: "unlimited"
                    overrides:
                      storage_quota_mb: 500
                      rate_read: "100mb/m"
        "400":
          description: Invalid pubkey format
        "401":
          description: Missing or invalid admin password
        "404":
          description: User not found

    patch:
      tags: [Admin]
      summary: Update user quota
      description: |
        Partially update per-user quota overrides.
        - Absent fields: keep existing value
        - `null`: reset to system default
        - `"unlimited"`: no limit
        - Value: explicit custom limit
      operationId: patchUserQuota
      servers:
        - url: http://localhost:6288
      security:
        - adminPassword: []
      parameters:
        - $ref: "#/components/parameters/AdminPubkey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserQuotaPatch"
            examples:
              setCustom:
                value:
                  storage_quota_mb: 500
                  rate_read: "100mb/m"
              resetToDefault:
                value:
                  storage_quota_mb: null
                  rate_read: null
                  rate_write: null
              setUnlimited:
                value:
                  rate_read: "unlimited"
      responses:
        "200":
          description: Quota updated
        "400":
          description: Invalid pubkey format
        "401":
          description: Missing or invalid admin password
        "404":
          description: User not found
        "422":
          description: Invalid quota format
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Grant-based bearer token from `POST /auth/grant/session`.

    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: |
        Deprecated. In practice, the cookie name is the user's z-base-32
        public key and the value is the session secret. The `name` field here
        is a placeholder; the actual name varies per user.

    adminPassword:
      type: apiKey
      in: header
      name: X-Admin-Password
      description: Admin server password.

    adminBasicAuth:
      type: http
      scheme: basic
      description: Basic auth for WebDAV endpoints. Username `admin`, password is the admin password.
  parameters:
...
    AdminPubkey:
      name: pubkey
      in: path
      required: true
      description: User's z-base-32 encoded public key.
      schema:
        type: string
  schemas:
...
    UserQuota:
      type: object
      description: |
        Quota configuration. All fields are optional.
        - Absent or `null`: use system default
        - `"unlimited"`: no limit
        - Value: explicit limit
      properties:
        storage_quota_mb:
          description: Storage quota in megabytes, `"unlimited"`, or `null` for default.
          oneOf:
            - type: integer
            - type: string
              enum: [unlimited]
        rate_read:
          description: "Read bandwidth limit (e.g. `\"200mb/m\"`, `\"unlimited\"`, or `null`)."
          type: string
        rate_write:
          description: "Write bandwidth limit (e.g. `\"100mb/m\"`, `\"unlimited\"`, or `null`)."
          type: string
        rate_read_burst:
          description: |
            Burst override for read rate limit, in the rate's natural unit
            (MB for "…mb/s", KB for "…kb/s"). `null` or absent = burst equals rate (default).
          type: integer
          format: int32
          minimum: 1
        rate_write_burst:
          description: |
            Burst override for write rate limit, in the rate's natural unit
            (MB for "…mb/s", KB for "…kb/s"). `null` or absent = burst equals rate (default).
          type: integer
          format: int32
          minimum: 1
        allowed_write_paths:
          description: |
            Restrict which paths a user can write to.
            - `null` or absent: unrestricted (all paths allowed) — default
            - `[]`: read-only (no writes allowed)
            - `["/pub/tokens/", "/pub/paykit/"]`: only these prefix paths
          type: array
          items:
            type: string
    UserQuotaPatch:
      type: object
      description: |
        Partial quota update. Only fields present in the body are changed.
        - Absent: keep existing value (no change)
        - `null`: reset to system default
        - `"unlimited"`: no limit
        - Value: explicit custom limit
      properties:
        storage_quota_mb:
          description: Storage quota in MB, `"unlimited"`, or `null` to reset.
          oneOf:
            - type: integer
            - type: string
              enum: [unlimited]
            - type: "null"
        rate_read:
          description: "Read bandwidth limit string, `\"unlimited\"`, or `null` to reset."
          type: ["string", "null"]
        rate_write:
          description: "Write bandwidth limit string, `\"unlimited\"`, or `null` to reset."
          type: ["string", "null"]
        rate_read_burst:
          description: "Burst for read rate limit (in rate's natural unit), or `null` to reset to default."
          type: ["integer", "null"]
          format: int32
          minimum: 1
        rate_write_burst:
          description: "Burst for write rate limit (in rate's natural unit), or `null` to reset to default."
          type: ["integer", "null"]
          format: int32
          minimum: 1
        allowed_write_paths:
          description: |
            Allowed write paths. Absent = keep existing, `null` = reset to unrestricted,
            `[]` = read-only, `["/pub/tokens/"]` = only these prefix paths.
          type: ["array", "null"]
          items:
            type: string
    UserQuotaResponse:
      type: object
      required: [effective, overrides]
      properties:
        effective:
          description: Overrides merged with system defaults. All fields are always present.
          allOf:
            - $ref: "#/components/schemas/UserQuota"
        overrides:
          description: Only per-user customizations. Fields using the system default are omitted.
          allOf:
            - $ref: "#/components/schemas/UserQuota"